@@ -1,16 +1,15 @@
-import { EmptyShoppingBagIcon, HeartIcon, ShoppingBagIcon } from '@/components'
-import { ButtonLink } from '@/components/Links'
+import { EmptyBagIcon, HeartIcon, BagIcon, ButtonLink } from '@/components'
+
 import { PATHS } from '@/constants'
-import { useWishList } from '@/context/global-context'
+import { useWishList } from '@/context/wish-list-context'
 import { EmptyItemsSection } from '@/features/product'
 import { Divider, Grid, Typography } from '@mui/material'
-import { forwardRef } from 'react'
 
-export const EmptyBagSection = forwardRef((props, ref) => {
+export const EmptyBagSection = (props) => {
   const wishList = useWishList()
 
   return (
-    <EmptyItemsSection Icon={EmptyShoppingBagIcon} {...props} ref={ref}>
+    <EmptyItemsSection Icon={EmptyBagIcon} {...props}>
       <Typography variant="h6" component="h2" children="Your bag is empty" fontWeight={500} />
       <Typography
         variant="body2"
@@ -20,11 +19,11 @@ export const EmptyBagSection = forwardRef((props, ref) => {
         <Grid container direction="column" alignItems="center" rowGap={3}>
           <ButtonLink href={PATHS.WISH_LIST} children="Go to Wish List" endIcon={<HeartIcon />} />
           <Divider flexItem variant="middle" children="or" />
-          <ButtonLink href={PATHS.HOME} children="Start Shopping" endIcon={<ShoppingBagIcon />} />
+          <ButtonLink href={PATHS.HOME} children="Start Shopping" endIcon={<BagIcon />} />
         </Grid>
       ) : (
-        <ButtonLink href={PATHS.HOME} children="Start Shopping" endIcon={<ShoppingBagIcon />} />
+        <ButtonLink href={PATHS.HOME} children="Start Shopping" endIcon={<BagIcon />} />
       )}
     </EmptyItemsSection>
   )
-})
+}
