@@ -1,4 +1,4 @@
-export { FreeDeliveryAlert } from './FreeDeliveryAlert'
+export { FreeShippingAlert } from './FreeShippingAlert'
 export { PromotionCodeAccordion } from './PromotionCodeAccordion'
 export { EmptyBagSection } from './EmptyBagSection'
 export { ViewedProductsSection } from './ViewedProductsSection'
