@@ -31,13 +31,13 @@ export {
   Add as AddIcon,
   ExpandMore as AccordionOpenIcon,
   InfoOutlined as InformationIcon,
-  LocalShippingOutlined as DeliveryIcon,
+  LocalShippingOutlined as ShippingIcon,
   ArrowForward as ArrowForwardIcon,
   DeleteOutlined as DeleteIcon,
   Clear as ClearIcon,
   Approval as ApproveIcon,
-  ShoppingBag as ShoppingBagIcon,
-  ShoppingBagOutlined as EmptyShoppingBagIcon,
+  ShoppingBag as BagIcon,
+  ShoppingBagOutlined as EmptyBagIcon,
   ReceiptOutlined as ReceiptIcon,
   AccessTime as ClockIcon,
   PublishedWithChanges as ValidateIcon,
