@@ -6,8 +6,8 @@ import { Link, HeartIcon, IconButton, CoverImage } from '../../../../components'
 import { useEffectOnMount } from '@/hooks'
 
 import styles from './styles'
-import { isHoverable } from '@/theming'
-import { useWishList } from '@/context/global-context'
+import { isHoverable } from '@/theme'
+import { useWishList } from '@/context/wish-list-context'
 import { wait } from '@/utils/helpers'
 import { useSnackbar } from '@/context/snackbar-context'
 import { ProductPricing } from '../ProductPricing'
